+++ src/Controller/AdminRedirectController.php
@@ -2,7 +2,10 @@
 
 namespace Scop\PlatformRedirecter\Controller;
 
+use Scop\PlatformRedirecter\MessageQueue\Message\CheckRedirectMessage;
 use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\Dbal\Common\RepositoryIterator;
+use Shopware\Core\Framework\DataAbstractionLayer\Dbal\Common\SalesChannelRepositoryIterator;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\Log\Package;
@@ -10,6 +13,7 @@ use Shopware\Core\Framework\Store\InAppPurchase;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Routing\Attribute\Route;
 
 #[Route(defaults: ['_routeScope' => ['api']])]
@@ -18,7 +22,7 @@ class AdminRedirectController extends AbstractController
 {
     const inAppPurchaseId = 'my-iap-identifier';
 
-    public function __construct(private readonly EntityRepository $redirectRepository, private readonly InAppPurchase $inAppPurchase)
+    public function __construct(private readonly EntityRepository $redirectRepository, private readonly InAppPurchase $inAppPurchase, private readonly MessageBusInterface $bus)
     {
     }
 
@@ -26,29 +30,30 @@ class AdminRedirectController extends AbstractController
     public function checkRedirects(Request $request, Criteria $criteria, Context $context): JsonResponse
     {
         if ($this->inAppPurchase->isActive('ScopPlatformRedirecter', self::inAppPurchaseId) === false) {
-           /* return new JsonResponse([
-                'total' => 0,
-                'broken' => 0,
-                'error' => 'in_app_purchase_invalid',
-            ]);*/
+            /* return new JsonResponse([
+                 'total' => 0,
+                 'broken' => 0,
+                 'error' => 'in_app_purchase_invalid',
+             ]);*/
         }
 
         $criteria = new Criteria();
-        $redirects = $this->redirectRepository->search($criteria, $context);
+        $iterator = new RepositoryIterator($this->redirectRepository, $context, $criteria);
 
-        foreach ($redirects as $redirect) {
+        $total = $iterator->getTotal();
 
+        if ($total > 0) {
+            while ($redirectItem = $iterator->fetch()) {
+                foreach ($redirectItem->getEntities() as $redirect) {
+                    $this->bus->dispatch(new CheckRedirectMessage($redirect));
+                }
+            }
         }
 
-        dd($redirects);
-
-
-        $broken = [];
         $total = 0;
 
         return new JsonResponse([
             'total' => $total,
-            'broken' => $broken,
             'error' => ''
         ]);
     }