+++ CHANGELOG.md
@@ -111,3 +111,4 @@
 
 - Added search function
 - Added primary key
+- Added broken redirect check